@@ -10,16 +10,16 @@
     "prebuild:css": "rimraf lib/css && mkdir -p lib/css",
     "build:css": "node scripts/buildCSS.js",
     "clean": "rimraf lib esm",
-    "precover": "rimraf coverage",
+    "precover": "rimraf coverage && npm run react",
     "cover": "cross-env NODE_ENV=test node --max-old-space-size=2048 $(which nyc) npm run mocha test",
     "lint": "eslint --ext .js,.jsx src test",
     "mocha": "mocha ./test/_helpers",
     "storybook:uninstall": "npm uninstall --no-save @storybook/react && rimraf node_modules/@storybook node_modules/react-modal node_modules/react-dom-factories",
-    "react:clean": "npm run storybook:uninstall && npm i --no-save ajv ajv-keywords && npm uninstall --no-save react react-dom react-addons-test-utils react-test-renderer && rimraf node_modules/react-test-renderer node_modules/react && npm prune",
-    "react:14": "rimraf node_modules/.bin/npm && npm run react:clean && npm i --no-save react@0.14 react-dom@0.14 react-addons-test-utils@0.14 enzyme-adapter-react-14 && npm run storybook:uninstall && npm prune && npm i --no-save enzyme-adapter-react-14",
-    "react:15": "rimraf node_modules/.bin/npm && npm run react:clean && npm i --no-save react@15 react-dom@15 react-addons-test-utils@15 react-test-renderer@15 && npm run storybook:uninstall && npm prune",
+    "react": "enzyme-adapter-react-install 16",
     "pretest": "npm run --silent lint",
+    "pretests-only": "npm run react",
     "tests-only": "npm run mocha --silent test",
+    "pretests-karma": "npm run react",
     "tests-karma": "karma start",
     "test": "npm run tests-only",
     "prestorybook": "npm run build",
@@ -70,10 +71,10 @@
     "chai": "^4.1.2",
     "clean-css": "^4.1.9",
     "coveralls": "^2.13.3",
-    "cross-env": "^5.1.0",
-    "enzyme": "^3.1.0",
-    "enzyme-adapter-react-15": "^1.0.2",
-    "eslint": "^4.9.0",
+    "cross-env": "^5.1.1",
+    "enzyme": "^3.2.0",
+    "enzyme-adapter-react-helper": "^1.0.3",
+    "eslint": "^4.11.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
@@ -91,22 +92,19 @@
     "karma-webpack": "^2.0.5",
     "mocha": "^3.5.3",
     "mocha-wrap": "^2.1.1",
-    "moment": "^2.19.1",
+    "moment": "^2.19.2",
     "moment-jalaali": "^0.7.2",
     "node-sass": "^4.5.3",
     "nyc": "^11.2.1",
     "raw-loader": "^0.5.1",
-    "react": "^0.14 || ^15.5.4",
-    "react-addons-shallow-compare": "^0.14 || ^15.5.2",
-    "react-addons-test-utils": "^0.14 || ^15.5.1",
-    "react-dom": "^0.14 || ^15.5.4",
-    "react-test-renderer": "^15.6.1",
+    "react": "^0.14 || ^15.5.4 || ^16.1.1",
+    "react-dom": "^0.14 || ^15.5.4 || ^16.1.1",
     "react-with-styles-interface-aphrodite": "^3.1.1",
-    "react-with-styles-interface-css-compiler": "^1.0.1",
+    "react-with-styles-interface-css-compiler": "^1.1.0",
     "rimraf": "^2.6.2",
     "safe-publish-latest": "^1.1.1",
     "sass-loader": "^6.0.6",
-    "sinon": "^4.0.1",
+    "sinon": "^4.1.2",
     "sinon-sandbox": "^1.0.2",
     "style-loader": "^0.19.0",
     "webpack": "^2.6.1"
@@ -126,9 +125,8 @@
   },
   "peerDependencies": {
     "moment": "^2.18.1",
-    "react": ">=0.14",
-    "react-dom": ">=0.14",
-    "react-addons-shallow-compare": ">=0.14"
+    "react": "^0.14 || ^15.5.4 || ^16.1.1",
+    "react-dom": "^0.14 || ^15.5.4 || ^16.1.1"
   },
   "greenkeeper": {
     "ignore": [
