@@ -1,9 +1,3 @@
-const configure = require('enzyme').configure;
-let Adapter;
-try {
-  Adapter = require('enzyme-adapter-react-15');
-} catch (e) {
-  Adapter = require('enzyme-adapter-react-14');
-}
+import configure from 'enzyme-adapter-react-helper';
 
-configure({ adapter: new Adapter(), disableLifecycleMethods: true });
+configure({ disableLifecycleMethods: true });
